move funcs into button container
