+++ src/components/ButtonContainer.js
@@ -1,14 +1,48 @@
 import React from 'react'
-import { FilePicker } from 'react-file-picker'
 import Spinner from './Spinner'
+import { FilePicker } from 'react-file-picker'
+import { importCSV } from '../functions/csv'
 
 function ButtonContainer({
 	createRecord,
 	cleanUpInvalidRecords,
 	exportAndDownloadCSV,
-	handleCsvImport,
 	loading,
+	setLoading,
+	updateRecords,
+	db,
 }) {
+	async function importCSVText(text) {
+		setLoading(true)
+		try {
+			console.log(text)
+			await importCSV(text, db)
+			console.log(db)
+			updateRecords()
+		} catch (err) {
+			console.log('error', err)
+			alert('Error when importing database: ' + err)
+			setLoading(false)
+		}
+		return setLoading(false)
+	}
+
+	async function handleCsvImport(importCsvFile) {
+		if (!window.confirm(`Importing may overwrite data. Continue?`)) return
+		if (importCsvFile.text) {
+			let text = await importCsvFile.text()
+			importCSVText(text)
+		} else {
+			// For safari that lacks the .text() call
+			var fileReader = new FileReader()
+			fileReader.addEventListener('loadend', (e) => {
+				var text = e.srcElement.result
+				importCSVText(text)
+			})
+			fileReader.readAsText(importCsvFile)
+		}
+	}
+
 	return (
 		<div className="button-container">
 			<div className="buttons">
