+++ src/functions/csv.js
@@ -66,7 +66,6 @@ function exportCSV(codebook, records) {
 
 async function importCSV(text, db) {
 	let promise = new Promise((resolve, reject) => {
-		debugger
 		csvParse(
 			text,
 			{
